Complete the work needed for the library rename
