+++ sxbp/solve.c
