+++ sxbp/solve.h
